Update format of celery tasks
This commit updates the format for celery tasks. Celery_instance
externalizes the app from tasks.py in order to support specifying tasks
within a separate file. To use the task a user needs to specify the
module in import_py. Without this change the method to manage tasks is
more fragile and generates a giant tasks.py file.

Service files are moved to etc. Amqp typo is fixed. Run directory
location is fixed. Systemd for service management is specified.
Placement of mapped variable data is fixed for compatibility with puppet
4 agents. A reference to app_dir is fixed in flower. Permissions are set
to correct default for systemd.